Move TestSuite and TestCase to parser
